feat(commands.ts): refactor fix function to support both UI and non-UI mode
The fix function in commands.ts has been refactored to support both UI and non-UI mode. In UI mode, the function now takes a screenshot of the target file and uses it for generating and fixing code. In non-UI mode, the function generates and fixes code without taking a screenshot.

The changes were made to improve the flexibility of the fix function and allow it to be used in different scenarios.

feat(commands.ts): add support for process.env.PORT environment variable to be able to run app on a configurable port

The server.ts file has been updated to add support for the process.env.PORT environment variable. This allows the app to be run on a configurable port, which can be set using the PORT environment variable.

The changes were made to improve the flexibility of the app and allow it to be easily deployed to different environments with different port configurations.

fix(server.ts): change port variable case from lowercase port to uppercase PORT to improve semantics

The server.ts file has been updated to change the case of the port variable from lowercase "port" to uppercase "PORT". This change improves the semantics of the variable and makes it more clear that it represents the port number.

The change was made to improve the readability and maintainability of the code.